refactor: 여행 계획 삭제 api 위치 이동, 이름 변경 및 로직 개선 	modified:   src/apis/services/article/reader/type.ts 	modified:   src/apis/services/article/writer/fetch.ts 	modified:   src/apis/services/article/writer/type.ts 	modified:   src/apis/services/article/writer/useService.ts 	modified:   src/apis/services/articleSchedule/writer/fetch.ts 	modified:   src/apis/services/articleSchedule/writer/type.ts 	modified:   src/apis/services/articleSchedule/writer/useService.ts
